requirements.txt
djorm-ext-filtered-contenttypes==0.2
django-pgjson==0.2.3
